Move io related test under io.personium.test.io.* package
